--- packages/design-system/src/components/button.stories.ts
@@ -1,8 +1,0 @@
-import { html } from "lit-html";
-import "../../dist/components/button.js";
-
-export default {
-	title: "Button",
-};
-
-export const Primary = () => html`<in-button>hello world</in-button>`;
