+++ website/src/routes/+page.svelte
@@ -1,2 +1,4 @@
 <script lang="ts">
 </script>
+
+<in-button name="Samuel" />
